feat: cache support + $USE_LATEST_CHROME
save chrome to ~/downloads folder that is cached
downloading is managed by the environmental variable $USE_LATEST_CHROME (if true, it will install the latest chrome, if anything else, it does nothing)

+++ use_chrome_stable_version.sh
@@ -1,10 +1,17 @@
 #!/bin/bash
 
-function use_chrome_stable_version {
-  curl -L -o google-chrome.deb https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb;
+use_chrome_stable_version() {
+  echo 'Upgrading the current Chrome version:';
+  google-chrome --version
+  cd $HOME/downloads
+  if ! test -f "google-chrome.deb"
+  then
+    curl -L -o google-chrome.deb https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb;
+  fi
   sudo dpkg -i google-chrome.deb;
   sudo sed -i 's|HERE/chrome\"|HERE/chrome\" --disable-setuid-sandbox|g' /opt/google/chrome/google-chrome;
-  rm google-chrome.deb
+  echo 'The latest Chrome version is:';
+  google-chrome --version
 }
 
 use_chrome_stable_version
